+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.6.0] - 2023-05-21
+
 ### Added
 - Multiple tag manager container IDs can be used (#14)
 - Option for requiring cookie consent (#20)
@@ -66,7 +68,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial release
 
 
-[Unreleased]: https://github.com/brotkrueml/typo3-matomo-integration/compare/v1.5.1...HEAD
+[Unreleased]: https://github.com/brotkrueml/typo3-matomo-integration/compare/v1.6.0...HEAD
+[1.6.0]: https://github.com/brotkrueml/typo3-matomo-integration/compare/v1.5.1...v1.6.0
 [1.5.1]: https://github.com/brotkrueml/typo3-matomo-integration/compare/v1.5.0...v1.5.1
 [1.5.0]: https://github.com/brotkrueml/typo3-matomo-integration/compare/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/brotkrueml/typo3-matomo-integration/compare/v1.3.2...v1.4.0